+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - PostgreSQL, MS SQL, Oracle: Do not highlight a valid date or time value as too long
 - PostgreSQL: Alter the trigger function from table structure
 - PostgreSQL: Do not highlight an array value as too long
+- PostgreSQL: Fix renaming a column and making it a foreign key at the same time
 - MS SQL: Export and alter the precision of datetime2, time and datetimeoffset
 - MS SQL: Keep fractional seconds and the time zone offset when saving a row with the sqlsrv extension
 - MS SQL Dblib: Display uniqueidentifier (bug #1339)